requirements.txt
numpy>=1.24.0
sounddevice>=0.4.6
python-rtmidi>=1.5.0
